Make an import safe validator for Filter.filter
.. by not touching anything in models.py.
