+++ src/argus/incident/constants.py
@@ -1,0 +1,5 @@
+# Prevent import loops
+# DO NOT import anything here, ever
+
+INCIDENT_LEVELS = (1, 2, 3, 4, 5)
+INCIDENT_LEVEL_CHOICES = zip(INCIDENT_LEVELS, map(str, INCIDENT_LEVELS))
