+++ src/argus/notificationprofile/validators.py
@@ -4,6 +4,7 @@ from typing import Union
 from rest_framework import serializers
 
 from .constants import DEPRECATED_FILTER_NAMES
+from .primitive_serializers import FilterBlobSerializer
 
 
 def validate_filter_string(value: Union[str, dict]):
@@ -33,3 +34,14 @@ def validate_filter_string(value: Union[str, dict]):
         errors.append(serializers.ValidationError(f"Unknown fieldname(s) in filterstring: {pp_unknown}", "unknown"))
     if errors:
         raise serializers.ValidationError(errors)
+
+
+def validate_jsonfilter(value: dict):
+    if not isinstance(value, dict):
+        raise serializers.ValidationError("Filter is not a dict")
+    if not value:
+        return True
+    serializer = FilterBlobSerializer(value)
+    if serializer.is_valid():
+        return True
+    raise serializers.ValidationError("Filter is not valid")
